+++ src/Properties/Resources.Designer.cs
@@ -20,7 +20,7 @@ namespace LibHeifSharp.Properties {
     // class via a tool like ResGen or Visual Studio.
     // To add or remove a member, edit your .ResX file then rerun ResGen
     // with the /str option, or rebuild your VS project.
-    [global::System.CodeDom.Compiler.GeneratedCodeAttribute("System.Resources.Tools.StronglyTypedResourceBuilder", "16.0.0.0")]
+    [global::System.CodeDom.Compiler.GeneratedCodeAttribute("System.Resources.Tools.StronglyTypedResourceBuilder", "17.0.0.0")]
     [global::System.Diagnostics.DebuggerNonUserCodeAttribute()]
     [global::System.Runtime.CompilerServices.CompilerGeneratedAttribute()]
     internal class Resources {
@@ -76,6 +76,15 @@ namespace LibHeifSharp.Properties {
         internal static string CannotGetAllAuxillaryImages {
             get {
                 return ResourceManager.GetString("CannotGetAllAuxillaryImages", resourceCulture);
+            }
+        }
+        
+        /// <summary>
+        ///   Looks up a localized string similar to Could not get all of the decoding warnings..
+        /// </summary>
+        internal static string CannotGetAllDecodingWarnings {
+            get {
+                return ResourceManager.GetString("CannotGetAllDecodingWarnings", resourceCulture);
             }
         }